Sorts functions moved to another .c file
